requirements.txt
Flask==3.1.2
flask-login==0.6.3
flask-cors==4.0.1
PyJWT==2.8.0
db==0.1.1
db-sqlite3==0.0.1